Move functions to __init__.py
